+++ runtime/vm/os_linux.cc
@@ -410,10 +410,6 @@ class JitDumpCodeObserver : public CodeObserver {
 
 #endif  // !PRODUCT
 
-const char* OS::Name() {
-  return "linux";
-}
-
 intptr_t OS::ProcessId() {
   return static_cast<intptr_t>(getpid());
 }